+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+### HEAD
+* Balance: Toned down mimic in endless mode
+
 ### v1.3.0
 * Added cards: 1 red, 1 blue
 * Added relics: Dentures, Gift Box